add JWT authentication and project routes

+++ src/controllers/controller.js
@@ -6,9 +6,11 @@ require('dotenv').config();
 //POST 
 const shortenUrl = async(req,res) => {
 const{originalUrl} = req.body;
+const userId = req.user.userId;
+
 try{
     const[existing] =await db.query(
-        'SELECT * FROM urls WHERE original_url =?',[originalUrl]
+        'SELECT * FROM urls WHERE original_url =? AND user_id = ? ',[originalUrl, userId]
     );
     if(existing.length>0){
         const url = existing[0];
@@ -21,8 +23,8 @@ try{
 
     const shortCode= nanoid(6);
     await db.query(
-        'INSERT INTO urls (original_url, short_code) VALUES (?,?)',
-        [originalUrl,shortCode]
+        'INSERT INTO urls (original_url, short_code, user_id) VALUES (?,?,?)',
+        [originalUrl, shortCode, userId]
     );
 
     res.status(201).json({
@@ -41,16 +43,17 @@ try{
 
  const redirectUrl = async(req, res) =>{
     const {shortCode} = req.params;
+    const userId = req.user.userId;
     try{
         const[rows] = await db.query(
-            'SELECT * FROM urls WHERE short_code = ?',[shortCode]
+            'SELECT * FROM urls WHERE short_code = ? AND user_id = ? ',[shortCode,userId]
         );
         if(rows.length === 0)
             return res.status(404).json({error: 'SHORT URLs not found'});
 
         //increment click_count
         await db.query(
-            'UPDATE urls SET click_count=click_count+1 WHERE short_code =? ',[shortCode]
+            'UPDATE urls SET click_count=click_count+1 WHERE short_code = ? AND user_id = ?',[shortCode, userId]
         );
         res.redirect(302, rows[0].original_url)
     }
@@ -63,10 +66,13 @@ try{
  // getAllUrls
   
 const getAllUrls = async(req,res) =>{
+
+    const userId = req.user.userId;
     
     try{
         const[rows] = await db.query(
-            'SELECT id, original_url, short_code, click_count, created_at FROM urls ORDER BY created_at DESC' 
+            'SELECT id, original_url, short_code, click_count, created_at FROM urls WHERE user_id = ? ORDER BY created_at DESC', 
+            [userId]
         );
 
         res.json(rows);
@@ -80,10 +86,11 @@ const getAllUrls = async(req,res) =>{
 
 const deleteUrl = async(req,res) => {
     const{shortCode} = req.params;
+    const userId = req.user.userId;
 
     try{
         const [result] = await db.query(
-            'DELETE FROM urls WHERE short_code =?',[shortCode]
+            'DELETE FROM urls WHERE short_code =? AND user_id = ?',[shortCode, userId]
         );
         if(result.affectedRows === 0)
             return res.status(404).json({error: 'Short URLs not found'});
